requirements.txt
pynput
Pillow
pyinstaller